requirements.txt
Flask==3.0.0
python-docx==1.1.0
python-pptx==0.6.23
Pillow==10.1.0
pandas==2.1.3
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0
reportlab==4.0.7
PyPDF2==3.0.1
pytesseract==0.3.10
openpyxl==3.1.2
Werkzeug==3.0.1
Jinja2==3.1.2
python-dateutil==2.8.2
